+++ src/main/java/com/unerp/dto/product/ProductResponse.java
@@ -1,13 +1,13 @@
 package com.unerp.dto.product;
 
 import java.time.LocalDate;
-
+import java.math.BigDecimal;
 public record ProductResponse(
     Integer id,
     String name,
     String description,
     Integer stock,
-    Double price,
+    BigDecimal price,
     String batch,
     LocalDate expirationDate,
     Integer supplierId
